requirements.txt
amqp==2.1.4
appdirs==1.4.3
asn1crypto==0.22.0
attrs==17.2.0
Automat==0.6.0
autopep8==1.3.1
Babel==2.4.0
backports.ssl-match-hostname==3.5.0.1
billiard==3.5.0.2
celery==4.0.2
certifi==2017.1.23
cffi==1.10.0
click==6.7
constantly==15.1.0
cryptography==2.0.2
cssselect==1.0.1
cycler==0.10.0
dask==0.14.0
ddtrace==0.8.0
decorator==4.0.11
Django==1.10.6
djangorestframework==3.5.4
editdistance==0.3.1
enum34==1.1.6
Flask==0.11
Flask-SQLAlchemy==2.2
Flask-SSLify==0.1.5
flower==0.9.1
funcsigs==1.0.2
functools32==3.2.3.post2
futures==3.0.5
google-api-python-client==1.6.2
h5py==2.6.0
httplib2==0.10.3
hyperlink==17.2.1
idna==2.5
incremental==17.5.0
ipaddress==1.0.18
itsdangerous==0.24
Jinja2==2.9.5
jsonpickle==0.9.4
Keras==1.2.2
kombu==4.0.2
lxml==3.8.0
MarkupSafe==1.0
matplotlib==2.0.0
mmh3==2.3.1
mock==2.0.0
msgpack-python==0.4.8
MySQL-python==1.2.5
mysqlclient==1.3.10
networkx==1.11
nltk==3.2.2
numpy==1.12.1
oauth2client==4.0.0
olefile==0.44
opencv-python==3.2.0.6
packaging==16.8
pandas==0.20.1
parsel==1.2.0
pbr==3.0.0
pdfminer==20140328
pdftables==0.0.4
Pillow==4.0.0
protobuf==3.3.0
pyasn1==0.2.3
pyasn1-modules==0.0.8
pycodestyle==2.3.1
pycparser==2.17
PyDispatcher==2.0.5
pyOpenSSL==17.2.0
pyparsing==2.2.0
pyPdf==1.13
python-dateutil==2.6.0
python-magic==0.4.13
pytz==2016.10
PyYAML==3.12
queuelib==1.4.2
requests==2.13.0
rsa==3.4.2
scikit-image==0.12.3
scikit-learn==0.18.1
scipy==0.18.1
Scrapy==1.4.0
selenium==3.4.3
service-identity==17.0.0
six==1.10.0
SQLAlchemy==1.1.11
subprocess32==3.2.7
swalign==0.3.3
tabula-py==0.9.0
tabulate==0.7.7
tensorflow==1.1.0
termcolor==1.1.0
Theano==0.8.2
toolz==0.8.2
tornado==4.2
tqdm==4.11.2
Twisted==17.5.0
uritemplate==3.0.0
uWSGI==2.0.14
vine==1.1.3
w3lib==1.17.0
Werkzeug==0.12.1
wrapt==1.10.10
xgboost==0.6a2
zope.interface==4.4.2
